(fish) Refactor theme switcher
Use a consistent order in steps of each branch. Restart Kitty when
running to prevent errors when run from a different terminal.

+++ .config/fish/functions/theme.fish
@@ -1,15 +1,19 @@
 function theme --description "Switches theme to light or dark mode" --argument-names mode
     if test $mode = light
+        echo light >"$HOME/.theme"
         fish_config theme choose "Rosé Pine Dawn"
         cat "$HOME/.config/kitty/themes/rose-pine-dawn.conf" >"$HOME/.config/kitty/current-theme.conf"
-        kitty @ set-colors -a "$HOME/.config/kitty/themes/rose-pine-dawn.conf"
-        kill -SIGUSR1 $KITTY_PID
-        echo light >"$HOME/.theme"
+
+        if set -q KITTY_PID
+            kill -SIGUSR1 $KITTY_PID
+        end
     else if test $mode = dark
-        fish_config theme choose "Rosé Pine Moon"
-        kitty @ set-colors -a "$HOME/.config/kitty/themes/rose-pine-moon.conf"
-        cat "$HOME/.config/kitty/themes/rose-pine-moon.conf" >"$HOME/.config/kitty/current-theme.conf"
-        kill -SIGUSR1 $KITTY_PID
         echo dark >"$HOME/.theme"
+        fish_config theme choose "Rosé Pine Moon"
+        cat "$HOME/.config/kitty/themes/rose-pine-moon.conf" >"$HOME/.config/kitty/current-theme.conf"
+
+        if set -q KITTY_PID
+            kill -SIGUSR1 $KITTY_PID
+        end
     end
 end
